+++ NEWS.md
@@ -1,3 +1,8 @@
+# trajr 1.2.1
+
+* Altered handling of parameter `readcsvFn` to `TrajsBuild` to make it 
+  possible to use `readr::read_csv` without a wrapper function.
+
 # trajr 1.2.0
 
 * Added start.pt.cex parameter to function `lines.Trajectory`.